pyproject.toml
[project]
name = "mix_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.2",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "snowflake-connector-python>=3.15.0",
    "snowflake-snowpark-python>=1.32.0",
    "tabulate>=0.9.0",
]
